+++ src/mac/MacJoyStick.cpp
@@ -105,7 +105,7 @@ void MacJoyStick::capture()
 		{
 			case kIOHIDElementTypeInput_Button: {
 				std::vector<IOHIDElementCookie>::iterator buttonIt = std::find(mCookies.buttonCookies.begin(), mCookies.buttonCookies.end(), event.elementCookie);
-				int button										   = std::distance(mCookies.buttonCookies.begin(), buttonIt);
+				int button										   = (int)std::distance(mCookies.buttonCookies.begin(), buttonIt);
 				mState.mButtons[button]							   = (event.value == 1);
 
 				if(mBuffered && mListener)
@@ -121,7 +121,7 @@ void MacJoyStick::capture()
 				//TODO: It's an axis! - kind of - for gamepads - or should this be a pov?
 			case kIOHIDElementTypeInput_Axis:
 				std::map<IOHIDElementCookie, AxisInfo>::iterator axisIt = std::find_if(mCookies.axisCookies.begin(), mCookies.axisCookies.end(), FindAxisCookie(event.elementCookie));
-				int axis												= std::distance(mCookies.axisCookies.begin(), axisIt);
+				int axis												= (int)std::distance(mCookies.axisCookies.begin(), axisIt);
 
 				//Copied from LinuxJoyStickEvents.cpp, line 149
 				const AxisInfo& axisInfo = axisIt->second;
@@ -209,7 +209,7 @@ void MacJoyStick::_enumerateCookies()
 				continue;
 			if(!CFNumberGetValue((CFNumberRef)object, kCFNumberIntType, &number))
 				continue;
-			min = number;
+			min = (int)number;
 
 			//Get max
 			object = CFDictionaryGetValue(element,
@@ -218,7 +218,7 @@ void MacJoyStick::_enumerateCookies()
 				continue;
 			if(!CFNumberGetValue((CFNumberRef)object, kCFNumberIntType, &number))
 				continue;
-			max = number;
+			max = (int)number;
 
 			//Get usage page
 			object = CFDictionaryGetValue(element,
@@ -260,8 +260,8 @@ void MacJoyStick::_enumerateCookies()
 			}
 		}
 
-		mInfo->numButtons = mCookies.buttonCookies.size();
-		mInfo->numAxes	  = mCookies.axisCookies.size();
+		mInfo->numButtons = (int)mCookies.buttonCookies.size();
+		mInfo->numAxes	  = (int)mCookies.axisCookies.size();
 	}
 	else
 	{
